+++ Homework/Lesson5/task2/ICircle.php
@@ -2,5 +2,5 @@
 
 interface ICircle
 {
-    function circleArea(int $circumference);
+    function circleArea(float $circumference);
 }
